+++ include/hardware.h
@@ -1,0 +1,8 @@
+#ifndef _HARDWARE_H_
+#define _HARDWARE_H_
+
+void timer();
+char readkey();
+
+#endif /*! _HARDWARE_ */
+//Lort
