+++ include/spil.h
@@ -1,6 +1,7 @@
 #ifndef _SPIL_H_
 #define _SPIL_H_
 void runStriker(struct Striker * striker,char push);
-void spil()
+void spil();
 
 #endif /*! _SPIL_H_ */
+//Lort
